receiveValue with weak self closure
This allows store to be released before effect is complete, without creating a
retain cycle. Ordinarily store's lifetime is the lifetime of the app, so
this should not be an issue for single-store apps, but it should be done this
way regardless. Additionally, this will now avoid a possible retain cycle if
using multiple short-lived stores, such as one per component.