Receive fx publisher on main
Receive Fx on main thread. This does two important things:

First, SwiftUI requires that any state mutations that would change
views happen on the main thread. Receiving on main ensures that
all fx-driven state transitions happen on main, even if the
publisher is off-main-thread.

Second, if we didn't schedule receive on main, it would be possible
for publishers to complete immediately, causing receiveCompletion
to attempt to remove the publisher from `cancellables` before
it is added. By scheduling to receive publisher on main,
we force publisher to complete on next tick, ensuring that it
is always first added, then removed from `cancellables`.