Receive Fx on main thread (#6)
This does two important things:

First, SwiftUI requires that any state mutations that would change views happen on the main thread. Receiving on main ensures that all fx-driven state transitions happen on main, even if the publisher is off-main-thread.

Second, if we don’t schedule to receive on main, it is possible for publishers to complete immediately, causing receiveCompletion to attempt to remove the publisher from `cancellables` before it is added. This was happening with `Empty()` publishers returned by default when you do not specify `Update(fx:)`.

By scheduling to receive publisher on main, we force publisher to complete on next tick, ensuring that it is always first added, then removed from `cancellables`.

This is an alternative to the flag-based approach in https://github.com/gordonbrander/ObservableStore/blob/2022-03-12-fx-tests-2/Sources/ObservableStore/ObservableStore.swift#L175.

### More background

Cancellables will cancel themselves they are released (makes sense). This means you must keep cancellables alive for the duration of the lifetime of the publisher by storing them somewhere. In our case, we store them in a map of `[UUID: Cancellable]`.

So anyway, we remove them when publisher completes in the `receiveCompletion`. HOWEVER, it is possible for a publisher to IMMEDIATELY complete during the same tick. This causes the `receiveCompletion` code which removes the cancellable from the map to run BEFORE the cancellable is added to the map. The result is that these immediately-completing cancellables leak, in the sense that they build up and are never removed. Cancellables are very tiny, so we didn't notice the impact on memory, but it is bad hygiene.