Refactor everything around ModelProtcol
By defining a protocol for models, we can describe where to find the
update function and statically dispatch to it. This allows us to
simplify a number of APIs, and reduces the number of types we need to
pass down to Store/ViewStore.

ModelProtocol has Action and Environment as associated types, so...

- Update only needs one type parameter
- Store init does not need update passed in or stored as a closure
- Cursor does not need update passed in as a closure

Also rework the naming of action/model types to Store/ViewStore,
Model/ViewModel for clarity.