Synthesize update(state:actions:environment)
Also:
- Remove Update.pipe. Was never happy with this approach. Ineligant.
- Revert fx to be joined on main thread

Now that we have a way to immediately sequence actions, we no longer
need to run fx on same tick. Joining on main is my preference from an
API perspective because it has fewer footguns in implementation and use.

We join on main with a .default QoS, because fx should be async/never
block user interaction.