Allow initializing first update from closure (#30)
Introduce `init(create:environment:` convenience initializer.

Remove `init(update:action:environment:` convenience initializer.

This is an alternative to #29 for cases where we need to initialize store from an `Update`. Instead of introducing a `ModelFactoryProtocol`, we introduce a convenience initializer that takes a `create` closure.

An advantage here is that we do not have to create a `flags` type for customizing the model construction process. Instead, because we can choose which factory function is passed to `create`, we can create one-off closure factories that set the necessary properties on the model for a given view.

Additionally, this approach does not require us to specify the type signature of the Store, since Swift is able to infer the store type from the function's return type.

Fixes #28

Alternatives considered:

- #31
- #29.