Introduce Future extensions for bridging to async-await (#32)
Introduces extensions to `Future` that allow you to create a Future from an async closure. This gives us a nice bridge from async swift to Combine.

Also introduce other initializers for convenience, and to make dispatch un-ambiguous:

`Update(state:fx:transaction:)`
`Update(state:animation:)`
`Update(state:fx:animation:)`

Fixes #26 by bridging async/await to Combine.

## Notes on design and tradeoffs

This PR is an alternative to #27. As mentioned in https://github.com/subconsciousnetwork/ObservableStore/pull/27#issuecomment-1491899862, there are tradeoffs with a pure async/await-based approach to Fx

- We do want combine-like publisher-subscription semantics for things like https://github.com/subconsciousnetwork/subconscious/pull/473#issuecomment-1490878422. Therefore we would need combine publisher machinery in Store anyway.
- Additionally, we need a way to batch fx. If we pursue a pure async/await based approach, leaves us carrying an array of async closures on Update.
    - Why? Because we want a collection of Action-producing asynchronous tasks/processes, all which are all run in parallel. 
    - An async sequence is no good, because we need to resolve in parallel, not in sequence.
    - So that leaves us with an array of async closures, which are run in store.
    - However, this semantics essentially describes Publishers. Going with an array of async closures does little except lose us the affordances of Combine Publishers.

One motivation for #26 was to have an explicit semantics of cancellation for long-polling tasks, by forcing them to fold into state via the update function at each step. This is good practice for the normal case. However, there are cases where this is not desired, like https://github.com/subconsciousnetwork/subconscious/pull/473#issuecomment-1490878422, or for keyboard events. Also, it is possible to cancel long-polling publishers at their source through the mechanisms publishers provide. There is no need for Store to have cancellation semantics.

A reasonable approach forward is to stick with Publishers, but to make it very convenient to bridge from async swift to Publishers. In particular, we want to make it very convenient to produce `Future<Action, Never>`, since these are one-shot Fx that always succeed and never fail, so they have the desired property that you do one fx, which produces one action, then fold into state with update.

And so, we introduce an extension to `Future` which allows you to create a Future from an async closure.