Make update a protocol (#41)
- Store requires `Model.update` to return some type that conforms to `UpdateProtocol`.
- `Update<Model>` becomes a concrete implementation of `UpdateProtocol`
- Users of ObservableStore are free to define custom update types that carry additional information, e.g. state change metadata for UIKit UIViewControllers.
- Publish updates from `Store.updates` publisher. Updates are guaranteed to publish after state change.
    - Useful for driving UIViewControllers.

This is a non-breaking change for existing consumers of ObservableStore. The motivating reason for the change is that we wished to use ObservableStore to power complex UIViewControllers in Subconscious, where a diffing-based approach was not practicable. Allowing for the customization of `Update` lets individual implementations of Store provide additional metadata to controllers.

While working on this PR, I also discovered a simpler way to implement effects that does not require us to dynamically manage cancellables, by using `flatMap` to consume batches of fx.